+++ Cargo.toml
@@ -5,13 +5,13 @@ authors = ["Maxwell Flitton <maxwellflitton@gmail.com>"]
 edition = "2018"
 
 [dependencies]
-surrealdb = { version = "1.2.0",  features = ["protocol-ws", "protocol-http", "kv-mem"] }
-serde = "^1.0.164"
-futures = { version="0.3.30", features = ["executor"] }
+surrealdb = { version = "1.5.4",  features = ["protocol-ws", "protocol-http", "kv-mem"] }
+serde = "1.0.210"
+futures = { version = "0.3.30", features = ["executor"] }
 once_cell = "1.19.0"
-serde_json = "^1.0.97"
-crossbeam-channel = "^0.5"
-thiserror = "^1.0.43"
+serde_json = "1.0.128"
+crossbeam-channel = "0.5.13"
+thiserror = "1.0.63"
 
 [dev-dependencies]
 tokio = { version = "1.36.0", features = ["full"] }
@@ -32,7 +32,7 @@ version = "0.20"
 features = ["unstable-streams", "attributes", "tokio-runtime"]
 
 [dependencies.uuid]
-version = "1.3.4"
+version = "1.10.0"
 features = [
     "v4",                # Lets you generate random UUIDs
     "fast-rng",          # Use a faster (but still sufficiently random) RNG
