+++ src/operations/create/core.rs
@@ -5,7 +5,6 @@
 //! * Create a record in the database
 use serde_json::value::Value;
 use surrealdb::opt::Resource;
-use surrealdb::sql::Range;
 use crate::connection::interface::WrappedConnection;
 
 
@@ -25,31 +24,11 @@ pub async fn create(connection: WrappedConnection, table_name: String, data: Val
 }
 
 
-/// Delete all records, or a specific record
-///
-/// # Arguments
-/// * `connection_id` - The connection performing the operation on the database
-/// * `resource` - The resource to delete (can be a table or a range)
-/// 
-/// # Returns
-/// 
-pub async fn delete(connection: WrappedConnection, resource: String) -> Result<String, String> {
-    let response = match resource.parse::<Range>() {
-        Ok(range) => {
-            connection.connection.delete(Resource::from(range.tb)).range((range.beg, range.end))
-                                 .await.map_err(|e| e.to_string())?
-        }
-        Err(_) => connection.connection.delete(Resource::from(resource))
-                                       .await.map_err(|e| e.to_string())?
-    };
-    Ok(response.into_json().to_string())
-}
-
-
 #[cfg(test)]
 mod tests {
 
     use super::*;
+    use super::super::super::delete::core::delete;
     use crate::operations::query::core::query;
     use crate::connection::core::make_connection;
 	use tokio::runtime::Runtime;
